Move the FileExtension enum to raw_assets
