+++ src/io.rs
@@ -70,111 +70,6 @@ mod pcd;
 #[path = "io/3mf.rs"]
 mod three_mf;
 
-enum FileExtension {
-    #[cfg(feature = "gltf")]
-    Gltf,
-    #[cfg(feature = "obj")]
-    Obj,
-    #[cfg(feature = "obj")]
-    Mtl,
-    #[cfg(feature = "stl")]
-    Stl,
-    #[cfg(feature = "fbx")]
-    Fbx,
-    #[cfg(feature = "pcd")]
-    Pcd,
-    #[cfg(feature = "3mf")]
-    ThreeMf,
-}
-
-impl FileExtension {
-    fn guess(raw_assets: &RawAssets, path: &Path) -> Result<Self> {
-        match extension(path).as_str() {
-            "" => match raw_assets.get(path) {
-                Ok(bytes) => Self::detect_from_bytes(path, bytes),
-                Err(_) => Err(Error::FailedDeserialize(path.to_str().unwrap().to_string())),
-            },
-            "gltf" | "glb" => {
-                #[cfg(not(feature = "gltf"))]
-                return Err(Error::FeatureMissing("gltf".to_string()));
-                #[cfg(feature = "gltf")]
-                Ok(Self::Gltf)
-            }
-            "obj" => {
-                #[cfg(not(feature = "obj"))]
-                return Err(Error::FeatureMissing("obj".to_string()));
-                #[cfg(feature = "obj")]
-                Ok(Self::Obj)
-            }
-            "stl" => {
-                #[cfg(not(feature = "stl"))]
-                return Err(Error::FeatureMissing("stl".to_string()));
-                #[cfg(feature = "stl")]
-                Ok(Self::Stl)
-            }
-            "fbx" => {
-                #[cfg(not(feature = "fbx"))]
-                return Err(Error::FeatureMissing("fbx".to_string()));
-                #[cfg(feature = "fbx")]
-                Ok(Self::Fbx)
-            }
-            "pcd" => {
-                #[cfg(not(feature = "pcd"))]
-                return Err(Error::FeatureMissing("pcd".to_string()));
-                #[cfg(feature = "pcd")]
-                Ok(Self::Pcd)
-            }
-            "mtl" => {
-                #[cfg(not(feature = "obj"))]
-                return Err(Error::FeatureMissing("obj".to_string()));
-                #[cfg(feature = "obj")]
-                Ok(Self::Mtl)
-            }
-            "3mf" => {
-                #[cfg(not(feature = "3mf"))]
-                return Err(Error::FeatureMissing("3mf".to_string()));
-                #[cfg(feature = "3mf")]
-                Ok(Self::ThreeMf)
-            }
-            _ => Err(Error::FailedDeserialize(path.to_str().unwrap().to_string())),
-        }
-    }
-
-    fn detect_from_bytes(path: &Path, bytes: &[u8]) -> Result<Self> {
-        if bytes.starts_with(b"glTF") {
-            #[cfg(not(feature = "gltf"))]
-            return Err(Error::FeatureMissing("gltf".to_string()));
-            #[cfg(feature = "gltf")]
-            return Ok(Self::Gltf);
-        }
-        if bytes.starts_with(b"Kaydara FBX Binary") {
-            #[cfg(not(feature = "fbx"))]
-            return Err(Error::FeatureMissing("fbx".to_string()));
-            #[cfg(feature = "fbx")]
-            return Ok(Self::Fbx);
-        }
-        if bytes.starts_with(&[0x50, 0x4B, 0x03, 0x04]) {
-            #[cfg(not(feature = "3mf"))]
-            return Err(Error::FeatureMissing("3mf".to_string()));
-            #[cfg(feature = "3mf")]
-            return Ok(Self::ThreeMf);
-        }
-        if bytes.starts_with(b"# .PCD") || bytes.starts_with(b"VERSION") {
-            #[cfg(not(feature = "pcd"))]
-            return Err(Error::FeatureMissing("pcd".to_string()));
-            #[cfg(feature = "pcd")]
-            return Ok(Self::Pcd);
-        }
-        if bytes.starts_with(b"solid ") {
-            #[cfg(not(feature = "stl"))]
-            return Err(Error::FeatureMissing("stl".to_string()));
-            #[cfg(feature = "stl")]
-            return Ok(Self::Stl);
-        }
-        Err(Error::FailedDeserialize(path.to_str().unwrap().to_string()))
-    }
-}
-
 ///
 /// Deserialize a single file from raw bytes. The key is used to determine the type of file.
 ///
@@ -239,6 +134,7 @@ pub trait Serialize: Sized {
 }
 
 use crate::{Error, Geometry, Result};
+use raw_assets::{extension, FileExtension};
 use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 
@@ -300,7 +196,7 @@ impl Serialize for crate::Texture2D {
 impl Deserialize for crate::Scene {
     fn deserialize(path: impl AsRef<Path>, raw_assets: &mut RawAssets) -> Result<Self> {
         let path = raw_assets.match_path(path.as_ref())?;
-        let ext = FileExtension::guess(raw_assets, &path)?;
+        let ext = raw_assets.guess_extension(&path)?;
         match ext {
             #[cfg(feature = "gltf")]
             FileExtension::Gltf => gltf::deserialize_gltf(raw_assets, &path),
@@ -422,7 +318,7 @@ fn dependencies(raw_assets: &RawAssets) -> Vec<PathBuf> {
     #[allow(unused_mut)]
     let mut dependencies = HashSet::new();
     for (path, _) in raw_assets.iter() {
-        match FileExtension::guess(raw_assets, path) {
+        match raw_assets.guess_extension(path) {
             #[cfg(feature = "gltf")]
             Ok(FileExtension::Gltf) => {
                 dependencies.extend(gltf::dependencies(raw_assets, path));
@@ -446,12 +342,6 @@ fn dependencies(raw_assets: &RawAssets) -> Vec<PathBuf> {
         .into_iter()
         .filter(|d| !raw_assets.contains_key(d))
         .collect()
-}
-
-fn extension(path: &Path) -> String {
-    path.extension()
-        .map(|e| e.to_str().unwrap().to_ascii_lowercase())
-        .unwrap_or("".to_string())
 }
 
 fn is_data_url(path: &Path) -> bool {
