+++ src/io/raw_assets.rs
@@ -5,6 +5,65 @@ use crate::{
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 
+pub(crate) enum FileExtension {
+    #[cfg(feature = "gltf")]
+    Gltf,
+    #[cfg(feature = "obj")]
+    Obj,
+    #[cfg(feature = "obj")]
+    Mtl,
+    #[cfg(feature = "stl")]
+    Stl,
+    #[cfg(feature = "fbx")]
+    Fbx,
+    #[cfg(feature = "pcd")]
+    Pcd,
+    #[cfg(feature = "3mf")]
+    ThreeMf,
+}
+
+impl FileExtension {
+    fn detect_from_bytes(path: &Path, bytes: &[u8]) -> Result<Self> {
+        if bytes.starts_with(b"glTF") {
+            #[cfg(not(feature = "gltf"))]
+            return Err(Error::FeatureMissing("gltf".to_string()));
+            #[cfg(feature = "gltf")]
+            return Ok(Self::Gltf);
+        }
+        if bytes.starts_with(b"Kaydara FBX Binary") {
+            #[cfg(not(feature = "fbx"))]
+            return Err(Error::FeatureMissing("fbx".to_string()));
+            #[cfg(feature = "fbx")]
+            return Ok(Self::Fbx);
+        }
+        if bytes.starts_with(&[0x50, 0x4B, 0x03, 0x04]) {
+            #[cfg(not(feature = "3mf"))]
+            return Err(Error::FeatureMissing("3mf".to_string()));
+            #[cfg(feature = "3mf")]
+            return Ok(Self::ThreeMf);
+        }
+        if bytes.starts_with(b"# .PCD") || bytes.starts_with(b"VERSION") {
+            #[cfg(not(feature = "pcd"))]
+            return Err(Error::FeatureMissing("pcd".to_string()));
+            #[cfg(feature = "pcd")]
+            return Ok(Self::Pcd);
+        }
+        if bytes.starts_with(b"solid ") {
+            #[cfg(not(feature = "stl"))]
+            return Err(Error::FeatureMissing("stl".to_string()));
+            #[cfg(feature = "stl")]
+            return Ok(Self::Stl);
+        }
+        Err(Error::FailedDeserialize(path.to_str().unwrap().to_string()))
+    }
+}
+
+pub(crate) fn extension(path: &Path) -> String {
+    path.extension()
+        .map(|e| e.to_str().unwrap().to_ascii_lowercase())
+        .unwrap_or("".to_string())
+}
+
 ///
 /// Contains raw assets which are usually generated by either the [load](crate::io::load)/[load_async](crate::io::load_async) functions or the [Serialize::serialize](crate::io::Serialize::serialize) function.
 /// Can also be constructed manually using the [RawAssets::new] and [RawAssets::insert].
@@ -61,6 +120,58 @@ impl RawAssets {
     ///
     pub fn get(&self, path: impl AsRef<Path>) -> Result<&[u8]> {
         Ok(self.0.get(&self.match_path(path.as_ref())?).unwrap())
+    }
+
+    pub(crate) fn guess_extension(&self, path: &Path) -> Result<FileExtension> {
+        match extension(path).as_str() {
+            "" => match self.get(path) {
+                Ok(bytes) => FileExtension::detect_from_bytes(path, bytes),
+                Err(_) => Err(Error::FailedDeserialize(path.to_str().unwrap().to_string())),
+            },
+            "gltf" | "glb" => {
+                #[cfg(not(feature = "gltf"))]
+                return Err(Error::FeatureMissing("gltf".to_string()));
+                #[cfg(feature = "gltf")]
+                Ok(FileExtension::Gltf)
+            }
+            "obj" => {
+                #[cfg(not(feature = "obj"))]
+                return Err(Error::FeatureMissing("obj".to_string()));
+                #[cfg(feature = "obj")]
+                Ok(FileExtension::Obj)
+            }
+            "stl" => {
+                #[cfg(not(feature = "stl"))]
+                return Err(Error::FeatureMissing("stl".to_string()));
+                #[cfg(feature = "stl")]
+                Ok(FileExtension::Stl)
+            }
+            "fbx" => {
+                #[cfg(not(feature = "fbx"))]
+                return Err(Error::FeatureMissing("fbx".to_string()));
+                #[cfg(feature = "fbx")]
+                Ok(FileExtension::Fbx)
+            }
+            "pcd" => {
+                #[cfg(not(feature = "pcd"))]
+                return Err(Error::FeatureMissing("pcd".to_string()));
+                #[cfg(feature = "pcd")]
+                Ok(FileExtension::Pcd)
+            }
+            "mtl" => {
+                #[cfg(not(feature = "obj"))]
+                return Err(Error::FeatureMissing("obj".to_string()));
+                #[cfg(feature = "obj")]
+                Ok(FileExtension::Mtl)
+            }
+            "3mf" => {
+                #[cfg(not(feature = "3mf"))]
+                return Err(Error::FeatureMissing("3mf".to_string()));
+                #[cfg(feature = "3mf")]
+                Ok(FileExtension::ThreeMf)
+            }
+            _ => Err(Error::FailedDeserialize(path.to_str().unwrap().to_string())),
+        }
     }
 
     pub(crate) fn match_path(&self, path: &Path) -> Result<PathBuf> {
